Exclude some AnnotationConfigUtils beans
This commit related to 2524218b987f0906f8604693e9bf7615eb8cc312 excludes those
AnnotationConfigUtils beans because they perform too much reflection for
scanning purpose (for example on configuration classes):
 - AUTOWIRED_ANNOTATION_PROCESSOR_BEAN_NAME
 - COMMON_ANNOTATION_PROCESSOR_BEAN_NAME
 - EVENT_LISTENER_PROCESSOR_BEAN_NAME
 - EVENT_LISTENER_FACTORY_BEAN_NAME

Instead of using directly those, we probably need to find a way to perform
the scanning part on AOT and the injection at runtime with related extension
points to add the required reflection config.

As a consequence, for now, autowiring of fields/setters as well as event
listeners are probably broken.